+++ tests/testthat/test-earth.R
@@ -119,7 +119,7 @@ test_that("Tests with parsnip returns no alert", {
   expect_false(
     tidypredict_test(
       parsnip::fit(
-        parsnip::set_engine(parsnip::mars(), "earth"),
+        parsnip::set_engine(parsnip::mars(mode = "regression"), "earth"),
         survived ~ age + sibsp,
         data = etitanic
       ),
@@ -129,7 +129,7 @@ test_that("Tests with parsnip returns no alert", {
   expect_false(
     tidypredict_test(
       parsnip::fit(
-        parsnip::set_engine(parsnip::mars(), "earth"),
+        parsnip::set_engine(parsnip::mars(mode = "regression"), "earth"),
         age ~ sibsp + parch,
         data = etitanic
       ),
@@ -139,7 +139,7 @@ test_that("Tests with parsnip returns no alert", {
   expect_false(
     tidypredict_test(
       parsnip::fit(
-        parsnip::set_engine(parsnip::mars(prod_degree = 2), "earth"),
+        parsnip::set_engine(parsnip::mars(prod_degree = 2, mode = "regression"), "earth"),
         age ~ sibsp + parch,
         data = etitanic
       ),
@@ -149,7 +149,7 @@ test_that("Tests with parsnip returns no alert", {
   expect_false(
     tidypredict_test(
       parsnip::fit(
-        parsnip::set_engine(parsnip::mars(prod_degree = 3), "earth"),
+        parsnip::set_engine(parsnip::mars(prod_degree = 3, mode = "regression"), "earth"),
         age ~ sibsp + parch,
         data = etitanic
       ),