+++ tests/testthat/test_xgboost.R
@@ -44,7 +44,7 @@ test_that("Confirm SQL function returns a query", {
 
 context("xgboost-parsnip")
 test_that("Predictions are correct for different objectives", {
-  m <- parsnip::fit(parsnip::set_engine(parsnip::boost_tree(), "xgboost"), am ~ ., data = mtcars)
+  m <- parsnip::fit(parsnip::set_engine(parsnip::boost_tree(mode = "regression"), "xgboost"), am ~ ., data = mtcars)
   expect_false(tidypredict_test(m, df = mtcars)$alert)
 })