+++ NEWS.md
@@ -2,6 +2,7 @@
 
 - Adds support for categorical predictors in `partykit`
 
+- Fixes `parsnip` tests to meet standards of new CRAN version
 
 # tidypredict 0.4.2